requirements.txt
django==3.2.2
argon2-cffi==20.1.0
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
pillow==8.2.0
psycopg2==2.8.6
stripe==2.58.0
boto3==1.17.97
django-storages==1.11.1
django-taggit==1.5.1
django-allauth==0.44.0
django-rest-auth==0.9.5
django-crispy-forms==1.12.0
crispy-tailwind==0.5.0
django-ckeditor==6.1.0
django-tailwind==2.2.0
markdown==3.3.4
gunicorn==20.1.0
whitenoise==5.2.0
channels==3.0.4
django-redis==5.0.0
sqlparse==0.3.1
asgiref==3.4.1
pytz==2020.5
mutagen==1.45.1
django-celery-beat==2.2.1
celery==5.1.2
python-dateutil==2.8.1
